x86-64: create fewer fdefns
Reduce the prominence of fdefns by doing away with them for symbols.
Lisp calls are emitted through a linkage table which is conceptually
similar to the alien linkage table, either as
  CALL [rip-n] ; from immobile text space
or
  MOV tbl, [thread+n] ; load table base
  CALL [tbl+n] ; from dynamic space

fdefns are still used for non-symbol function names primarily as a
place to store the linkage table index.

See internals-notes/fdefn-removal for further details.

This builds through make-target-contrib. Haven't run tests yet.