Add full paths for gmp and mpfr on m1/m2
Homebrew on m1/m2 devices installs native packages to /opt/homebrew
instead of /usr/local. /opt/homebrew is not in the default search path
dlopen(3). The default search path for dlopen can be changed with
DYLD_LIBRARY_PATH, but doing so can be difficult because its partially
restricted by sip.